[Behat] Extract ProductTaxonomyTrait from product pages
